Move container setup to the dedicated module
